# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Login2FARequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      remember_me:
        type: boolean
    required:
    - email
    - password
    type: object
  models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      two_fa:
        description: Is 2FA turned on for this user?
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Verify2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
host: localhost:8088
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a web app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Web App API
  version: "1.0"
paths:
  /api/2fa/disable:
    post:
      description: Turn off 2FA for the current user
      responses:
        "200":
          description: 2FA disabled successfully
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Disable 2FA
      tags:
      - auth
  /api/2fa/setup:
    post:
      description: Generate a 2FA secret and QR code URI
      produces:
      - application/json
      responses:
        "200":
          description: Secret and QR URI
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Setup 2FA
      tags:
      - auth
  /api/2fa/verify:
    post:
      consumes:
      - application/json
      description: Confirm 2FA setup with a 6-digit code
      parameters:
      - description: 2FA Code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Verify2FARequest'
      responses:
        "200":
          description: 2FA enabled successfully
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Verify and Enable 2FA
      tags:
      - auth
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a message
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            type: string
      summary: Login user
      tags:
      - auth
  /api/login/2fa:
    post:
      consumes:
      - application/json
      description: Complete login process with 2FA code
      parameters:
      - description: Email and 2FA Code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Login2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login tokens
          schema:
            type: string
      summary: Login with 2FA
      tags:
      - auth
  /api/logout:
    post:
      description: Invalidate the user's refresh token
      responses:
        "200":
          description: Logged out successfully
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/profile:
    get:
      consumes:
      - application/json
      description: Get the profile of the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - auth
  /api/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New access token
          schema:
            type: string
      summary: Refresh access token
      tags:
      - auth
  /api/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            type: string
      summary: Register a new user
      tags:
      - auth
  /health:
    get:
      description: Check if the server is up and running
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health Check
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"